chore: move expandable ad, add translations
