+++ apps/web/src/components/AdPanel/Expandable/ExpandableAd.tsx
@@ -4,15 +4,15 @@ import { useRouter } from 'next/router'
 import { BodyText } from '../BodyText'
 import { AdButton } from '../Button'
 import { AdCard } from '../Card'
-import { ExpandableActions } from '../Expandable/ExpandableActions'
-import { ExpandableContent } from '../Expandable/ExpandableContent'
-import { ExpandableModal } from '../Expandable/ExpandableModal'
-import { useExpandableCard } from '../Expandable/useExpandableCard'
 import { FAQ } from '../FAQ'
 import { faqTypeByPage } from '../FAQ/config'
 import { useFaqConfig } from '../FAQ/useFaqConfig'
 import { Title } from '../Title'
 import { AdPlayerProps } from '../types'
+import { ExpandableActions } from './ExpandableActions'
+import { ExpandableContent } from './ExpandableContent'
+import { ExpandableModal } from './ExpandableModal'
+import { useExpandableCard } from './useExpandableCard'
 
 const ExpandedContent: React.FC = () => {
   const router = useRouter()
