+++ apps/web/src/components/AdPanel/Variations/ContentCountdown.tsx
@@ -1,3 +1,4 @@
+import { useTranslation } from '@pancakeswap/localization'
 import { BodyText } from '../BodyText'
 import { AdButton } from '../Button'
 import { AdCard } from '../Card'
@@ -6,9 +7,10 @@ import { AdPlayerProps } from '../types'
 import { getImageUrl } from '../utils'
 
 export const ContentCountdown = (props: AdPlayerProps) => {
+  const { t } = useTranslation()
   return (
     <AdCard imageUrl={getImageUrl('ad_clamm_options_trading.png')} {...props}>
-      <BodyText mb="8px">PancakeSwap Meetup in Turkey starts in:</BodyText>
+      <BodyText mb="8px">{t('PancakeSwap Meetup in Turkey starts in:')}</BodyText>
 
       <Countdown
         targetTime={1731997322}
@@ -19,7 +21,7 @@ export const ContentCountdown = (props: AdPlayerProps) => {
       />
 
       <AdButton variant="text" isExternalLink>
-        Learn More
+        {t('Learn More')}
       </AdButton>
     </AdCard>
   )
